Cleaned solution.cu
Cleaned and commented solution.cu. Color conversion, histogram calculation, and Otsu's Method are all in a working state. Also will output an image.

+++ canny/labs/Otsus_Method.h
@@ -3,5 +3,5 @@
 #include <chrono>
 using namespace std;
 
-void Histogram_Sequential(int* image, unsigned int* hist, int width, int height);
+void Histogram_Sequential(float* image, unsigned int* hist, int width, int height);
 double Otsu_Sequential(unsigned int* hist);
